refactor: Move providers into each feature directory
